pkg/k8s: refactor const.go to common directory
