+++ pkg/k8s/apis/cilium.io/utils/utils.go
@@ -17,7 +17,7 @@ package utils
 import (
 	"fmt"
 
-	k8sConst "github.com/cilium/cilium/pkg/k8s/apis/cilium.io"
+	k8sConst "github.com/cilium/cilium/pkg/k8s/apis"
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/logging"
 	"github.com/cilium/cilium/pkg/logging/logfields"
