+++ pkg/k8s/apis/cilium.io/v2/types.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/cilium/cilium/api/v1/models"
-	k8sConst "github.com/cilium/cilium/pkg/k8s/apis/cilium.io"
+	k8sConst "github.com/cilium/cilium/pkg/k8s/apis"
 	k8sUtils "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/utils"
 	"github.com/cilium/cilium/pkg/logging"
 	"github.com/cilium/cilium/pkg/logging/logfields"
