+++ pkg/k8s/apis/const.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package ciliumio
+package apis
 
 import (
 	"k8s.io/kubernetes/pkg/kubelet/types"
