+++ pkg/k8s/network_policy_v1beta.go
@@ -18,7 +18,7 @@ package k8s
 
 import (
 	"github.com/cilium/cilium/pkg/annotation"
-	k8sConst "github.com/cilium/cilium/pkg/k8s/apis/cilium.io"
+	k8sConst "github.com/cilium/cilium/pkg/k8s/apis"
 	k8sUtils "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/utils"
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/logging/logfields"
